# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Class/Course Organizer API Documentation",
    "version": "1.0.0",
    "description": "Documention for class and course organizer app. Perform CRUD operations on classes, courses, and students"
  },
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Student": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier for a student"
          },
          "name": {
            "type": "string",
            "description": "Full name of the student"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "Email address of the student"
          },
          "GPA": {
            "type": "number",
            "minimum": 0,
            "maximum": 4.5,
            "description": "Grade point average (GPA) of the student, ranging from 0 to 4.5"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/student": {
      "post": {
        "summary": "Creates a new student",
        "tags": [
          "Student"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Student"
              },
              "example": {
                "name": "Michael Scott",
                "email": "michaelscott@dmifflin.com",
                "GPA": 1.2
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The newly created student",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Student"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "status": "success",
                  "data": {
                    "id": "S123",
                    "name": "Michael Scott",
                    "email": "michaelscott@dmifflin.com",
                    "GPA": 1.2
                  },
                  "message": "Student created"
                }
              }
            }
          },
          "400": {
            "description": "Invalid inputs"
          },
          "403": {
            "description": "Unauthorized - Insufficient role"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "summary": "Gets all existing students",
        "tags": [
          "Student"
        ],
        "responses": {
          "200": {
            "description": "The list of students retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Student"
                      }
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "status": "success",
                  "data": [
                    {
                      "id": "S123",
                      "name": "Michael Scott",
                      "email": "michaelscott@dmifflin.com",
                      "GPA": 1.2
                    },
                    {
                      "id": "S124",
                      "name": "Dwight Schrute",
                      "email": "dschrute@dmifflin.com",
                      "GPA": 3.5
                    }
                  ],
                  "message": "Students retrieved"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized - Insufficient role"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/v1/student/{id}": {
      "get": {
        "summary": "Gets a student by id",
        "tags": [
          "Student"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "example": "S123"
            },
            "required": true,
            "description": "ID of the student to be retrieved"
          }
        ],
        "responses": {
          "200": {
            "description": "The student with the corresponding id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Student"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "status": "success",
                  "data": {
                    "id": "S123",
                    "name": "Michael Scott",
                    "email": "michaelscott@dmifflin.com",
                    "GPA": 1.2
                  },
                  "message": "Student retrieved"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized - Insufficient role"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "summary": "Updates an existing student",
        "tags": [
          "Student"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "example": "S123"
            },
            "required": true,
            "description": "ID of the student to be updated"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Student"
              },
              "example": {
                "name": "Michael Scott",
                "email": "michaelscott@dmifflin.com",
                "GPA": 1.2
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated student",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Student"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "status": "success",
                  "data": {
                    "id": "S123",
                    "name": "Michael Scott",
                    "email": "michaelscott@dmifflin.com",
                    "GPA": 1.2
                  },
                  "message": "Student updated"
                }
              }
            }
          },
          "400": {
            "description": "Invalid inputs"
          },
          "403": {
            "description": "Unauthorized - Insufficient role"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "summary": "Deletes an existing student",
        "tags": [
          "Student"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "example": "S123"
            },
            "required": true,
            "description": "ID of the student to be deleted"
          }
        ],
        "responses": {
          "200": {
            "description": "Student deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "status": "success",
                  "message": "Student deleted"
                }
              }
            }
          },
          "403": {
            "description": "Unauthorized - Insufficient role"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}